+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Cantrips would never display on spell sheet if setting was off
 - Copying an attack with linked ammunition to another actor made it behave strangely (#801)
 - Fix error on rolling concentration with no ability associated with the spellbook (#825)
+- `@abilities.<abl>.base` and `@abilities.<abl>.baseMod` weren't being updated properly (#824)
 
 ### Compendium